+++ CHANGELOG.md
@@ -10,6 +10,10 @@ We go to the next version after each release on [GNOME Shell Extensions website]
 
 - Blurry search entry on GNOME Shell themes with box-shadow.
 
+### Added
+
+- Panel world clock visiblity.
+
 ## [18.0.0 Roslin] - 2022-02-12
 
 ### Fixed